Blindness Trap Sabotage
The second sabotage - a blindness trap which hinders your map vision for 15 seconds - has been added. This exists alongside the existing stasis trap

The handling of sabotages has also been improved - most of this can now be found in the Sabotagable script.

One of my TDD tests was also updated to account for the fact multiple int values can now be randomly selected when determining the sabotage to apply.