Put Turret & Turret Bullet on Enemy Layer
Changes the turret and turret bullet prefabs to be on layer enemy, like the moving and jumping enemies. Now allows for spectators to completely ignore collisions with these objects.

Done as a separate commit so that this change can be rolled back more easily if this is found to break something.